Add token csrf pour remove item form cart

+++ src/Controller/CartController.php
@@ -5,8 +5,11 @@ namespace App\Controller;
 use App\Repository\ShoeRepository;
 use App\Service\CartService;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Attribute\Route;
+use Symfony\Component\Security\Csrf\CsrfToken;
+use Symfony\Component\Security\Csrf\CsrfTokenManagerInterface;
 
 class CartController extends AbstractController
 {
@@ -28,9 +31,12 @@ class CartController extends AbstractController
     }
 
     #[Route('/cart/{shoeId}/{waist}', name: 'app_cart_remove', methods: ['POST'])]
-    public function removeFormCart(int $shoeId, int $waist, CartService $cartService): Response
+    public function removeFormCart(int $shoeId, int $waist, CartService $cartService,Request $request, CsrfTokenManagerInterface $csrfTokenManager): Response
     {
-        $cartService->removeCart($shoeId, $waist);
+        $token = $request->request->get('_token');
+        if($csrfTokenManager->isTokenValid(new CsrfToken('remove_item', $token))) {
+            $cartService->removeCart($shoeId, $waist);
+        }
 
         return $this->redirectToRoute('app_cart');
     }
